working on csi format support
