+++ bamnostic/csi.py
@@ -1,6 +1,6 @@
 #!/user/bin/env python
 # -*- coding: utf-8 -*-
-"""BAI file parser
+"""CSI file parser
 
 .. include:: <isonum.txt>
 
@@ -10,14 +10,14 @@ This code is part of the bamnostic distribution and governed by its
 license.  Please see the LICENSE file that should have been included
 as part of this package.
 
-The Binary Alignment Map (BAM) format (defined at https://samtools.github.io/hts-specs/SAMv1.pdf)
-allows for indexing. When the user invokes a tool to index a given BAM file, a BAM index (BAI)
-file is created. Generally speaking, a BAI contains the all the virtual offsets of clusters of
-reads that are associated with specific subsections of the BAM file. When the BAI is produced,
+The Binary Alignment Map (CSI) format (defined at http://samtools.github.io/hts-specs/CSIv1.pdf)
+allows for indexing. When the user invokes a tool to index a given BAM file, a CSI index file
+can be created. Generally speaking, a CSI contains the all the virtual offsets of clusters of
+reads that are associated with specific subsections of the BAM file. When the CSI is produced,
 random access to the BAM is available.
 
-This script is for parsing the binary encoded BAI file, inherently making it human readable.
-Furthermore, it allows subsections of the BAI to be loaded into memory, reducing the memory
+This script is for parsing the binary encoded CSI file, inherently making it human readable.
+Furthermore, it allows subsections of the CSI to be loaded into memory, reducing the memory
 footprint and speeding up queries within a small number of references. Lastly, by parsing it
 as such, random access queries directly into the associated BAM file is available to other
 tools within bamnostic
@@ -42,7 +42,10 @@ if _PY_VERSION[0] == 2:
 else:
     from functools import lru_cache
 
+import bamnostic
+import bamnostic.bgzf as bgzf
 from bamnostic.utils import *
+from bamnostic.bai import *
 
 
 def format_warnings(message, category, filename, lineno, file=None, line=None):
@@ -71,29 +74,12 @@ class conditional_decorator(object):
         return self.decorator(func)
 
 
-# __slot__ classes for performant named indexing and future-proofing &readability
-RefIdx = namedtuple('RefIdx', ('start_offset', 'end_offset', 'n_bins'))
-
-
 class CsiBin(object):
     __slots__ = ['loffset', 'chunks']
 
     def __init__(self, *args):
         self.loffset, self.chunks = args
 
-
-class Chunk(object):
-    __slots__ = ['voffset_beg', 'voffset_end']
-
-    def __init__(self, handle):
-        self.voffset_beg, self.voffset_end = unpack_chunk(handle.read(16))
-
-
-class Ref(object):
-    __slots__ = ['bins', 'intervals', 'ref_id']
-
-    def __init__(self, *args):
-        self.bins, self.intervals, self.ref_id = args
 
 class RefCsi(object):
     __slots__ = ['bins', 'ref_id']
@@ -102,17 +88,7 @@ class RefCsi(object):
         self.bins, self.ref_id = args
 
 
-class Unmapped(object):
-    __slots__ = ['unmapped_beg', 'unmapped_end', 'n_mapped', 'n_unmapped']
-
-    def __init__(self, ubeg, uend, nmap, numap):
-        self.unmapped_beg = ubeg
-        self.unmapped_end = uend
-        self.n_mapped = nmap
-        self.n_unmapped = numap
-
-
-def reg2bin(rbeg, rend):
+def reg2bin_csi(rbeg, rend, min_shift=14, depth=5):
     """Finds the largest superset bin of region. Numeric values taken from hts-specs
 
     Args:
@@ -122,16 +98,23 @@ def reg2bin(rbeg, rend):
     Returns:
         (int): distinct bin ID for largest superset bin of region
     """
-    left_shift = 15
-    for i in range(14, 27, 3):
-        if rbeg >> i == (rend - 1) >> i:
-            return int(((1 << left_shift) - 1) / 7 + (rbeg >> i))
-        left_shift -= 3
-    else:
-        return 0
+
+    assert 0 <= rbeg <= rend, 'Invalid region {}, {}'.format(rbeg, rend)
+    total = ((1 << depth * 3) - 1) // 7
+    stop -= 1
+    while depth > 0:
+        
+        if start >> min_shift == stop >> min_shift:
+            return total + (start >> min_shift)
+        
+        min_shift += 3
+        total -= 1 << (depth * 3)
+        depth -= 1
+        
+    return 0
 
 
-def reg2bins(rbeg, rend):
+def reg2bins_csi(rbeg, rend, min_shift=14, depth=5):
     """Generates bin ids which overlap the specified region.
 
     Args:
@@ -145,29 +128,32 @@ def reg2bins(rbeg, rend):
         AssertionError (Exception): if the range is malformed or invalid
     """
     # Based off the algorithm presented in:
-    # https://samtools.github.io/hts-specs/SAMv1.pdf
+    # http://samtools.github.io/hts-specs/CSIv1.pdf
 
-    # Bin calculation constants.
-    BIN_ID_STARTS = (0, 1, 9, 73, 585, 4681)
+    assert 0 <= rbeg <= rend, 'Invalid region {}, {}'.format(rbeg, rend)
+    bins = []
+    shift = min_shift + depth * 3
+    stop -= 1
+    level = total = 0
+    
+    while level <= depth:
+        beg = total + (rbeg >> shift)
+        end = total + (rend >> shift)
+        
+        i = beg
+        while i <= end:
+            yield i
+            i += 1
+        shift -= 3
+        total += 1 << level * 3
+        level += 1
 
-    # Maximum range supported by specifications.
-    MAX_RNG = (2 ** 29) - 1
 
-    assert 0 <= rbeg <= rend <= MAX_RNG, 'Invalid region {}, {}'.format(rbeg, rend)
+class Csi(Bai):
+    """ This class defines the bam CSI index file object and its interface.
 
-    for start, shift in zip(BIN_ID_STARTS, range(29, 13, -3)):
-        i = rbeg >> shift if rbeg > 0 else 0
-        j = rend >> shift if rend < MAX_RNG else MAX_RNG >> shift
-
-        for bin_id_offset in range(i, j + 1):
-            yield start + bin_id_offset
-
-
-class Bai(object):
-    """ This class defines the bam index file object and its interface.
-
-    The purpose of this class is the binary parsing of the bam index file (BAI) associated with
-    a given bam file. When queried, the Bai object identifies the bins of data that overlap the
+    The purpose of this class is the binary parsing of the bam index file (CSI) associated with
+    a given bam file. When queried, the Csi object identifies the bins of data that overlap the
     requested region and directs which parts of the bam file contain it.
 
     Virtual offsets are processed using the following method:
@@ -175,23 +161,22 @@ class Bai(object):
     Position within uncompressed block = uoffset = virtual offset ^ (coffset << 16)
 
     Attributes:
-        _io (fileObject): opened BAI file object
-        _LINEAR_INDEX_WINDOW (int): constant of the linear interval window size
+        _io (fileObject): opened CSI file object
         _UNMAP_BIN (int): constant for bin ID of unmapped read stats
-        magic (bytes): first 4 bytes of file. Must be equal to b'BAI\x01'
-        n_refs (int): number of references in BAI
+        magic (bytes): first 4 bytes of file. Must be equal to b'CSI\x01'
+        n_refs (int): number of references in CSI
         unmapped (dict): dictionary of the unmapped read stats by each reference
         current_ref (None|dict): dictionary of the current reference loaded into memory.
                                 It contains the a dictionary of bin IDs and their respective
                                 chunks, and a list of linear intervals.
-        ref_indices (dict): dictionary of reference ids and their start/stop offsets within the BAI file
-        n_no_coord (None|int): if present in BAI, is the number of reads that have no coordinates
+        ref_indices (dict): dictionary of reference ids and their start/stop offsets within theCSI file
+        n_no_coord (None|int): if present inCSI, is the number of reads that have no coordinates
         _last_pos (int): used for indexing, the byte position of the file head.
 
     """
-    __slots__ = ['_io', '_LINEAR_INDEX_WINDOW', '_UNMAP_BIN', 'BAM_LIDX_SHIFT',
-                 'magic', 'n_refs', 'unmapped', 'current_ref', 'ref_indices',
-                 'n_no_coor', '_last_pos']
+    __slots__ = ['_io', '_min_shift', '_UNMAP_BIN', 'BAM_LIDX_SHIFT',
+                 '_magic', '_depth', 'aux', 'current_ref', 'ref_indices',
+                 'n_no_coor', '_last_pos', 'n_refs']
 
     def __init__(self, filename):
         """Initialization method
@@ -201,24 +186,32 @@ class Bai(object):
         out just the index information for that chromosome--not the whole genome.
 
         Args:
-            filename (str): '/path/to/bam_file' that automatically adds the '.bai' suffix
+            filename (str): '/path/to/bam_file' that automatically adds the '.csi' suffix
 
         Raises:
-            OSError (Exception): if the BAI file is not found or does not exist
-            AssertionError (Exception): if BAI magic is not found
+            OSError (Exception): if the CSI file is not found or does not exist
+            AssertionError (Exception): if CSI magic is not found
         """
         if os.path.isfile(filename):
-            self._io = open(filename, 'rb')
+            self._io = bgzf.BgzfReader(open(filename, 'rb'))
         else:
             raise OSError('{} not found. Please change check your path or index your BAM file'.format(filename))
 
         # Constant for linear index window size and unmapped bin id
-        self._LINEAR_INDEX_WINDOW = 16384
         self._UNMAP_BIN = 37450
-        self.BAM_LIDX_SHIFT = 14
 
-        self.magic, self.n_refs = unpack("<4sl", self._io)
-        assert self.magic == b'BAI\x01', 'Wrong BAI magic header'
+        self._magic, self._min_shift, self._depth, l_aux = unpack("<4s3i", self._io)
+        assert self._magic == b'CSI\x01', 'Wrong CSI magic header'
+
+        aux_nref = unpack('<{}Bi'.format(l_aux), self._io)
+        
+        if l_aux > 0:
+            self.aux = aux_nref[:-1]
+            self.n_ref = aux_nref[-1]
+        elif l_aux == 0:
+            self.n_refs = aux_nref
+        else:
+            raise IOError('CSI attribute l_aux must be >= 0')
 
         self.unmapped = {}
         self.current_ref = None
@@ -254,29 +247,6 @@ class Bai(object):
         chunks = [Chunk(self._io) for chunk in range(n_chunks)]
         return chunks
 
-    # TODO: Add a check for long reads and allow for skipping the linear index on queries
-    def get_ints(self, n_int):
-        """Unpacks `n_int` number of interval virtual offsets from stream
-
-        A linear interval is defined as a 16384 bp window along a given reference. The
-        value stored in the BAI is the virtual offset of the first read within
-        that given interval. This virtual offset is the byte offset (coffset) of the start of the
-        BGZF block that contains the beginning of the read and the byte offset (uoffset)
-        within the uncompressed data of the residing BGZF block to that first read.
-
-        Note: a caveat to using linear interval with long reads: A long read can
-        span multiple linear intervals. As such, the current encoding could potentially shift
-        the expected region of interest to the left more than expected.
-
-        Args:
-            n_int (int): number of intervals to unpack
-
-        Returns:
-            intervals (list): list of virtual offsets for `n_int` number of linear intervals
-        """
-        intervals = unpack('<{}Q'.format(n_int), self._io)
-        return intervals if type(intervals) != int else [intervals]
-
     def get_bins(self, n_bins, ref_id=None, idx=False):
         """Simple function that iteratively unpacks the data of a number (`n_bin`)
         of bins.
@@ -285,7 +255,7 @@ class Bai(object):
         bin, it deletages work to `self.get_chunks(n_chunks)`. A bin is
         comprised of 2 parts: 1) the distinct bin ID (within a given reference). If
         no reads are associated with that bin, it is left out of the indexing process,
-        and therefore not represented in the BAI file. Furthermore, while each bin
+        and therefore not represented in theCSI file. Furthermore, while each bin
         has a bin ID, the bin IDs are only distinct within a given reference. This means
         that 2 or more references can have the same bin IDs. These bin IDs are also
         not in any order as they are essentially a hash dump. Lastly, the only reserved
@@ -310,7 +280,7 @@ class Bai(object):
         bins = None if idx else {}
 
         for b in range(n_bins):
-            bin_id, n_chunks = unpack_bid_nchunk(self._io.read(8))
+            bin_id, loffset, n_chunks = unpack_bid_loffset_nchunk(self._io.read(12))
 
             if idx:
                 if bin_id == self._UNMAP_BIN:
@@ -322,7 +292,7 @@ class Bai(object):
                         self._io.seek(16 * n_chunks, 1)  # 16 = struct.calcsize('<2Q')
             else:
                 chunks = self.get_chunks(n_chunks)
-                bins[bin_id] = chunks
+                bins[bin_id] = CsiBin(loffset, chunks)
         else:
             return bins
 
@@ -332,27 +302,27 @@ class Bai(object):
     # @lru_cache(6)
     @conditional_decorator(lambda func: lru_cache(maxsize=6)(func), _PY_VERSION[0] == 2)
     def get_ref(self, ref_id=None, idx=False):
-        """Interatively unpacks all the bins, linear intervals, and chunks for a given reference
+        """Iteratively unpacks all the bins, linear intervals, and chunks for a given reference
 
         A reference is comprised of 2 things: 1) a series of bins that reference chunks of aligned
         reads that are grouped within that bin. 2) a series of virtual offsets of the first read of a
         16384 bp window along the given reference.
 
-        This function also serves to "index" the BAI file such that, if it is invoked by setting
-        `ids=True`, will do a single pass through the BAI file and saving the start and stop
+        This function also serves to "index" theCSI file such that, if it is invoked by setting
+        `ids=True`, will do a single pass through theCSI file and saving the start and stop
         offsets of each of the references. This is used for minimizing the memory footprint of
-        storing the BAI in memory. When queried against, the appropriate reference block will be
+        storing theCSI in memory. When queried against, the appropriate reference block will be
         loaded. Because of this constant loading, `functools.lru_cache` was applied to cache recently
         used reference blocks to speed up computation. It is assumed that when querying is done, most
         users are looking and just a few references at a time.
 
         Args:
-            ref_id (None|int): used for random access or indexing the BAI
-            idx (bool): Flag for setting whether or not to run an index of the BAI
+            ref_id (None|int): used for random access or indexing theCSI
+            idx (bool): Flag for setting whether or not to run an index of theCSI
 
         Returns:
             RefIdx: `namedtuple` containing the byte offsets of the reference start, stop, and number of bins
-            or
+            o
             Ref: `namedtuple` containing a dictionary of bins and list of linear intervals
 
         Raises:
@@ -373,18 +343,12 @@ class Bai(object):
         n_bins = unpack_int32L(self._io.read(4))[0]
         bins = self.get_bins(n_bins, ref_id, idx)
 
-        n_int = unpack_int32L(self._io.read(4))[0]
-        if idx:
-            self._io.seek(8 * n_int, 1)  # 8 = struct.calcsize('<Q')
-
-        ints = None if idx else self.get_ints(n_int)
-
         self._last_pos = self._io.tell()
 
         if idx:
             return RefIdx(ref_start, self._last_pos, n_bins)
         else:
-            return Ref(bins, ints, ref_id)
+            return RefCsi(bins, ref_id)
 
     def query(self, ref_id, start, stop=-1):
         """ Main query function for determining seek offset to BAM section that
@@ -409,67 +373,16 @@ class Bai(object):
         elif self.current_ref.ref_id != ref_id:
             self.current_ref = self.get_ref(ref_id)
 
-        # get linear index first
-        # how many windows do we need to go over
-        # because of floor div, we need to make it 0-based
-
-        reg_lin_idx = start >> self.BAM_LIDX_SHIFT
-
-        l_idx = reg_lin_idx if reg_lin_idx < len(self.current_ref.intervals) else -1
-        linear_offset = self.current_ref.intervals[l_idx]
-
-        for binID in reg2bins(start, stop):
+        for potential_bin in reg2bins_csi(start, stop, min_shift=self._min_shift, depth=self._depth):
             try:
-                bin_chunks = self.current_ref.bins[binID]
+                pbin = self.current_ref.bins[potential_bin]
             except KeyError:
                 continue
 
-            for chunk in bin_chunks:
-                if not linear_offset <= chunk.voffset_end:
-                    continue
-                else:
+            reg_lin_idx = start >> self._min_shift # BAI-style linear index should match with default loffset
+            
+            for chunk in pbin.chunks:
+                if pbin.loffset <= chunk.voffset_end:
                     return chunk.voffset_beg
-
-    def seek(self, offset=None, whence=0):
-        """Simple seek function for binary files
-
-        Args:
-            offset (None|int): byte offset from whence to move the file head to.
-            whence (int): 0 := from start of file, 1:= from current position, 2:= from end of file
-
-        Returns:
-            (int): new byte position of file head
-
-        Raise:
-            ValueError (Exception): if the offset is not an integer or is not provided
-        """
-        if isinstance(offset, (int, None)):
-            if offset is None:
-                raise ValueError('No offset provided')
-            else:
-                self._io.seek(offset, whence)
-                return self._io.tell()
-        else:
-            raise ValueError('offset must be an integer or None')
-
-    def read(self, size=-1):
-        """Simple read function for binary files
-
-        Args:
-            size (int): number of bytes to read in (default: -1 --whole file)
-
-        Returns:
-            (bytes): the number of bytes read from file
-        """
-        if size == 0:
-            return b''
-        else:
-            return self._io.read(size)
-
-    def tell(self):
-        """Simple tell function for reporting byte position of file head
-
-        Returns:
-            (int): byte position of file head
-        """
-        return self._io.tell()
+                else:
+                    continue
